// src/shared/generated/WorkspaceDisplayInfoRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StatePixelBoundsRecord } from "./StatePixelBoundsRecord";
import type { StateResolutionRecord } from "./StateResolutionRecord";

export type WorkspaceDisplayInfoRecord = { id: number, label: string, bounds: StatePixelBoundsRecord, resolution: StateResolutionRecord, scaleFactor: number, isPrimary: boolean, isInternal: boolean, };